timeweb/requirements.txt
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
cryptography==36.0.0
defusedxml==0.7.1
Django==3.2.12
django-allauth==0.49.0
django-cleanup==5.2.0
django-colorfield==0.4.3
django-contact-form==1.9
django-csp==3.7
django-multiselectfield==0.1.12
django-pwa @ git+https://github.com/ArhanChaudhary/django-pwa.git@16c0d555fefcec1f5d298f650cef9d0061c95343
django-storages==1.12.3
django-timezone-field==4.2.1
google-api-core==2.0.1
google-api-python-client==2.20.0
google-auth==2.0.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.0.0
google-cloud-storage==1.42.1
google-crc32c==1.1.5
google-resumable-media==2.0.1
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
oauthlib==3.1.1
packaging==21.0
Pillow==9.0.1
protobuf==3.17.3
psycopg2==2.9.1
psycopg2-binary==2.9.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.3.0
pyparsing==2.4.7
python-http-client==3.3.3
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sendgrid==6.9.1
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==2.0.3
uritemplate==3.0.1
urllib3==1.26.6
